requirements.txt
Flask
numpy
pillow
pytesseract
python-dotenv
waitress
flasgger
pyjwt